chore(fn): rename to mock
